kfp-pipelines/requirements.txt
kfp
google-cloud-aiplatform>=1.16.0
scikit-learn
pandas
numpy
joblib
